move go module to src/
