+++ src/runtime/cmd/ctrl/ctrl.go
@@ -41,13 +41,13 @@ import (
 	"syscall"
 	"time"
 
-	"go.corp.nvidia.com/osmo/pkg/args"
-	"go.corp.nvidia.com/osmo/pkg/common"
-	"go.corp.nvidia.com/osmo/pkg/data"
-	"go.corp.nvidia.com/osmo/pkg/messages"
-	"go.corp.nvidia.com/osmo/pkg/metrics"
-	"go.corp.nvidia.com/osmo/pkg/osmo_errors"
-	"go.corp.nvidia.com/osmo/pkg/rsync"
+	"go.corp.nvidia.com/osmo/runtime/pkg/args"
+	"go.corp.nvidia.com/osmo/runtime/pkg/common"
+	"go.corp.nvidia.com/osmo/runtime/pkg/data"
+	"go.corp.nvidia.com/osmo/runtime/pkg/messages"
+	"go.corp.nvidia.com/osmo/runtime/pkg/metrics"
+	"go.corp.nvidia.com/osmo/runtime/pkg/osmo_errors"
+	"go.corp.nvidia.com/osmo/runtime/pkg/rsync"
 
 	"github.com/gorilla/websocket"
 	"gopkg.in/yaml.v3"
