+++ src/runtime/pkg/args/ctrl_args.go
@@ -26,7 +26,7 @@ import (
 	"strings"
 	"time"
 
-	"go.corp.nvidia.com/osmo/pkg/common"
+	"go.corp.nvidia.com/osmo/runtime/pkg/common"
 )
 
 // Parse and process command line arguments
