+++ src/runtime/pkg/common/common.go
@@ -31,7 +31,7 @@ import (
 	"sort"
 	"sync"
 
-	"go.corp.nvidia.com/osmo/pkg/osmo_errors"
+	"go.corp.nvidia.com/osmo/runtime/pkg/osmo_errors"
 )
 
 type OsmoPort uint16
