+++ src/runtime/pkg/data/data.go
@@ -36,8 +36,8 @@ import (
 	"syscall"
 	"time"
 
-	"go.corp.nvidia.com/osmo/pkg/common"
-	"go.corp.nvidia.com/osmo/pkg/osmo_errors"
+	"go.corp.nvidia.com/osmo/runtime/pkg/common"
+	"go.corp.nvidia.com/osmo/runtime/pkg/osmo_errors"
 )
 
 var DataTimeout time.Duration = 10 * time.Minute
